+++ include/libopencm3/stm32/l1/adc.h
@@ -119,6 +119,7 @@ LGPL License Terms @ref lgpl_license
 #define ADC_CR1_PDI			(1 << 17)
 #define ADC_CR1_PDD			(1 << 16)
 
+#define ADC_CR1_AWDCH_MAX		26
 
 /* --- ADC_CR2 values ------------------------------------------------------- */
 /* SWSTART: */ /** Start conversion of regular channels. */
@@ -208,11 +209,18 @@ LGPL License Terms @ref lgpl_license
 /**@}*/
 
 #define ADC_SQR_MASK			0x1f
+#define ADC_SQR_MAX_CHANNELS_REGULAR	28 /* m+/h only, otherwise 27 */
 
 #define ADC_CCR_TSVREFE			(1 << 23)
 
 BEGIN_DECLS
-        // We will add these when we are ready...
+        /* L1 specific, or not fully inified adc routines */
+void adc_enable_temperature_sensor(void);
+void adc_disable_temperature_sensor(void);
+void adc_enable_external_trigger_regular(uint32_t adc, uint32_t trigger,
+					 uint32_t polarity);
+void adc_enable_external_trigger_injected(uint32_t adc, uint32_t trigger,
+					  uint32_t polarity);
 
 END_DECLS
