+++ lib/stm32/l1/Makefile
@@ -43,6 +43,7 @@ OBJS		+= usart_common_all.o usart_common_f124.o
 OBJS		+= exti_common_all.o
 OBJS		+= rcc_common_all.o
 OBJS		+= usb.o usb_control.o usb_standard.o usb_f103.o
+OBJS		+= adc.o adc_common_v1.o
 
 VPATH += ../../usb:../:../../cm3:../common
